requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
paddleocr==2.7.0.3
paddlepaddle==2.6.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
Pillow==10.2.0
pdf2image==1.17.0
camelot-py[cv]==0.11.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
